requirements.txt
# =============================================================================
# Real-Time Multi-Object Detection & Tracking System — Python Dependencies
# =============================================================================
# Install: pip install -r requirements.txt
# GPU-specific: install torch+cu121 from https://pytorch.org/get-started/locally/

# --- Core ---
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.1.0
onnxruntime-gpu>=1.16.0; sys_platform == "linux"
onnxruntime>=1.16.0; sys_platform == "win32"
# tensorrt>=8.6                   # Linux-only; install separately on target GPU host

# --- Tracking ---
numpy>=1.24.0
scipy>=1.11.0
lap>=0.4.0                        # linear assignment for ByteTrack
filterpy>=1.4.5                   # Kalman filter (DeepSORT)

# --- Video / RTSP ---
opencv-python-headless>=4.8.0
av>=11.0.0                        # optional: PyAV for GStreamer bridge

# --- Visualization ---
supervision>=0.18.0               # Roboflow supervision for annotation
Pillow>=10.0.0

# --- Evaluation ---
pycocotools>=2.0.7
motmetrics>=1.4.0

# --- Configuration & CLI ---
pyyaml>=6.0
click>=8.1.0

# --- Logging & Profiling ---
loguru>=0.7.0
psutil>=5.9.0
py-cpuinfo>=9.0.0
GPUtil>=1.4.0

# --- TensorBoard ---
tensorboard>=2.14.0

# --- Testing ---
pytest>=7.4.0
